10513: Move buildDesignSystem.ts
